requirements.txt
requests
Pillow
google-generativeai
python-dotenv
opencv-python